pyproject.toml
[tool.poetry]
name = "visionlog"
version = "0.1.0"
description = "A universal structured logging library for analytics tracking"
authors = ["Alan Szmyt <szmyty@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
loguru = "^0.7.0"
structlog = "^23.1.0"
orjson = "^3.8.1"
opentelemetry-api = "^1.15.0"
opentelemetry-sdk = "^1.15.0"
click = "^8.1.3"

[tool.poetry.scripts]
visionlog-cli = "visionlog.cli:log"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

